[improvement] DatetimePicker: split into two components (#3995)
